Add files via upload
# A Sudoku Game!

This program allows users to play **Sudoku** using the functions listed below:

1. `drawInputBoxes(first,second, x, y, w, rowNum, grid)`: This function takes in information for first, second, x, y, w, rowNum, and grid and returns a sudoku board with any given values filled in. The given values come from the grid entered (note that only values larger than 0 are written because 0 represents a missing value). The grid is a list of lists, with each nested list representing a row in the Sudoku board. First, second, x, y, w, and rowNum are integers. First and second affect the number of boxes in each row. X, y, and w represent the diminsions of each box. 
2. `lines(startX,startY,endX,endY)`: This function  draws a line given starting and ending points for that line. There are two starting points: startX and startY, both are given as integters or floats. startX gives the starting position on the x-axis abd startY gives the starting position on the y-axis. The same is true for the two ending points. 
3. `Setboard(grid)`: Given a grid, this function uses the drawInputBoxes function to set up a Sudoku Board. Setboard adds black lines in between each set of three columns to allow users to visualize the sets of three by three boxes that need each number 1-9. 
4. `inside(point, rectangle)`: This function is sourced from: https://stackoverflow.com/questions/39867464/adding-button-to-python-graphics-py-window. Inside takes in a rectangle and a point and returns True if the point is inside the rectangle and False otherwise.
5. `check(finishedgrid)`: This function takes in a finished grid and checks to see if the values the user entered match the values in the finished grid. If the user entered values all match the ones in the finished grid, the function prints a statement letting the user know. If the user entered an incorrect value, the function allows the user to either view the correct answer or keep guessing. 
6. `GiveAnswer(finishedgrid)`: This function takes in values for a finished suduku and uses the setboard(grid)function to draw a sudoku board with all the given values filled in. 

## Required Files:
For this Sudoku game to work, users need to have the graphics.py file. This file is used to create an illustration of the sudoku board and allow the user to enter guesses into boxes. The second required file is the SukokuGrids.py file which contains various Sudoku grids. Within the main file, a grid is randomly chosen out of the SudokuGrids.py file for the user to play. This allows the user to try out multiple Sudoku grids at random.

When running this project, one of the sudoku boards in the SukdokuGrids.py file is randomly chosen. Next the Setboard function is used to draw out the chosen sudoku game. Then, the check function is called, allowing the user to play sudoku by entering guesses. When the check() function is exited (by the user either entering a correct guess or indicating they want to see the answer), the user has the option to play again or not. If they chose to play again, the program loops back over all the previously mentioned items. Otherwise, the graphics window is closed.